Added scatterplot. Fixed x-axis update issue.
